refactor path location a bit to be more useful to loki. make graphreader easier to use if you already have a pt and dont need the hierarchy.  update graphtile to use size_t in place of int and to use c++ style casts instead of c style casts, also to use array operators instead of pointer arithmetic as its more obvious

+++ src/baldr/graphtileheader.cc
@@ -10,19 +10,19 @@ GraphTileHeader::GraphTileHeader()
       textlist_offset_(0) {
 }
 
-unsigned int GraphTileHeader::nodecount() const {
+size_t GraphTileHeader::nodecount() const {
   return nodecount_;
 }
 
-unsigned int GraphTileHeader::directededgecount() const {
+size_t GraphTileHeader::directededgecount() const {
   return directededgecount_;
 }
 
-unsigned int GraphTileHeader::edgeinfo_offset() const {
+size_t GraphTileHeader::edgeinfo_offset() const {
   return edgeinfo_offset_;
 }
 
-unsigned int GraphTileHeader::textlist_offset() const {
+size_t GraphTileHeader::textlist_offset() const {
   return textlist_offset_;
 }
